# api/openapi.json
{
  "components": {
    "schemas": {
...
      "ComputerOut": {
        "properties": {
          "acquired_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acquired Date"
          },
          "asset_id": {
            "title": "Asset Id",
            "type": "string"
          },
          "chassis": {
            "default": "",
            "title": "Chassis",
            "type": "string"
          },
          "condition": {
            "default": "",
            "title": "Condition",
            "type": "string"
          },
          "cpu": {
            "default": "",
            "title": "Cpu",
            "type": "string"
          },
          "disposed": {
            "default": false,
            "title": "Disposed",
            "type": "boolean"
          },
          "disposed_at": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disposed At"
          },
          "disposed_note": {
            "default": "",
            "title": "Disposed Note",
            "type": "string"
          },
          "drives": {
            "default": "",
            "title": "Drives",
            "type": "string"
          },
          "drives_note": {
            "default": "",
            "title": "Drives Note",
            "type": "string"
          },
          "image": {
            "default": "",
            "title": "Image",
            "type": "string"
          },
          "installed_ram": {
            "default": "",
            "title": "Installed Ram",
            "type": "string"
          },
          "installed_ram_kb": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Installed Ram Kb"
          },
          "installed_ram_note": {
            "default": "",
            "title": "Installed Ram Note",
            "type": "string"
          },
          "location": {
            "default": "",
            "title": "Location",
            "type": "string"
          },
          "location_path": {
            "default": "",
            "title": "Location Path",
            "type": "string"
          },
          "machine": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MachineOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "manufacturer": {
            "default": "",
            "title": "Manufacturer",
            "type": "string"
          },
          "model": {
            "default": "",
            "title": "Model",
            "type": "string"
          },
          "name": {
            "default": "",
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "default": "",
            "title": "Notes",
            "type": "string"
          },
          "os": {
            "default": "",
            "title": "Os",
            "type": "string"
          },
          "project": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project"
          },
          "serial": {
            "default": "",
            "title": "Serial",
            "type": "string"
          },
          "source": {
            "default": "",
            "title": "Source",
            "type": "string"
          },
          "summary": {
            "default": "",
            "title": "Summary",
            "type": "string"
          },
          "topbench": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topbench"
          },
          "url": {
            "default": "",
            "title": "Url",
            "type": "string"
          },
          "variant": {
            "default": "",
            "title": "Variant",
            "type": "string"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "required": [
          "asset_id"
        ],
        "title": "ComputerOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Retro Hardware Database API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/computers": {
      "get": {
        "operationId": "api_list_computers_api_computers_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ComputerOut"
                  },
                  "title": "Response Api List Computers Api Computers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Api List Computers",
        "tags": [
          "computers"
        ]
      },
      "post": {
        "operationId": "api_create_computer_api_computers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComputerCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComputerOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Create Computer",
        "tags": [
          "computers"
        ]
      }
    },
    "/api/computers/{aid}": {
      "delete": {
        "description": "Delete a computer. The parts inside it are unlinked, not deleted; its\nphotos, drive/memory rows and history go with it (see delete_computer). The\nAPI deletes any computer -- the disposed-only rule and the confirmation are\nthe GUI's, where a delete is a click rather than a deliberate request.",
        "operationId": "api_delete_computer_api_computers__aid__delete",
        "parameters": [
          {
            "in": "path",
            "name": "aid",
            "required": true,
            "schema": {
              "title": "Aid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Delete Computer",
        "tags": [
          "computers"
        ]
      },
      "get": {
        "operationId": "api_get_computer_api_computers__aid__get",
        "parameters": [
          {
            "in": "path",
            "name": "aid",
            "required": true,
            "schema": {
              "title": "Aid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComputerOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Get Computer",
        "tags": [
          "computers"
        ]
      },
      "patch": {
        "operationId": "api_update_computer_api_computers__aid__patch",
        "parameters": [
          {
            "in": "path",
            "name": "aid",
            "required": true,
            "schema": {
              "title": "Aid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComputerIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComputerOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Update Computer",
        "tags": [
          "computers"
        ]
      }
    },
...
    "/api/parts": {
      "get": {
        "operationId": "api_list_parts_api_parts_get",
        "parameters": [
          {
            "in": "query",
            "name": "computer_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Computer Id"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PartOut"
                  },
                  "title": "Response Api List Parts Api Parts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api List Parts",
        "tags": [
          "parts"
        ]
      },
      "post": {
        "operationId": "api_create_part_api_parts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Create Part",
        "tags": [
          "parts"
        ]
      }
    },
    "/api/parts/{aid}": {
      "delete": {
        "description": "Delete a part, with its typed spec rows, photos and history. Anything\nmounted on it is unlinked, not deleted.",
        "operationId": "api_delete_part_api_parts__aid__delete",
        "parameters": [
          {
            "in": "path",
            "name": "aid",
            "required": true,
            "schema": {
              "title": "Aid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Delete Part",
        "tags": [
          "parts"
        ]
      },
      "get": {
        "operationId": "api_get_part_api_parts__aid__get",
        "parameters": [
          {
            "in": "path",
            "name": "aid",
            "required": true,
            "schema": {
              "title": "Aid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Get Part",
        "tags": [
          "parts"
        ]
      },
      "patch": {
        "operationId": "api_update_part_api_parts__aid__patch",
        "parameters": [
          {
            "in": "path",
            "name": "aid",
            "required": true,
            "schema": {
              "title": "Aid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Update Part",
        "tags": [
          "parts"
        ]
      }
    },
...
  }
}